+++ client-core/src/main/java/com/influxdb/exceptions/InfluxException.java
@@ -50,6 +50,7 @@ import retrofit2.Response;
 public class InfluxException extends RuntimeException {
 
     private static final Logger LOG = Logger.getLogger(InfluxException.class.getName());
+    private static final String HTTP_STATUS_CODE_MESSAGE = "HTTP status code: %d; Message: %s";
 
     private final Response<?> response;
     private final String message;
@@ -88,7 +89,7 @@ public class InfluxException extends RuntimeException {
     }
 
     /**
-     * Gets the reference code unique to the error type. If the reference code is not present than return "0".
+     * Gets the reference code unique to the error type. If the reference code is not present then return "0".
      *
      * @return reference code unique to the error type
      */
@@ -108,7 +109,7 @@ public class InfluxException extends RuntimeException {
 
     /**
      * Gets the HTTP status code of the unsuccessful response.
-     * If the response is not present than return "0".
+     * If the response is not present then return "0".
      *
      * @return HTTP status code
      */
@@ -123,7 +124,7 @@ public class InfluxException extends RuntimeException {
 
     /**
      * Gets the HTTP headers from the unsuccessful response.
-     * If the response is not present than return empty {@code Map}.
+     * If the response is not present then return empty {@code Map}.
      *
      * @return HTTP headers
      */
@@ -155,6 +156,7 @@ public class InfluxException extends RuntimeException {
     @Nullable
     private String messageFromResponse() {
         if (response != null) {
+            int code = response.code();
             try {
                 ResponseBody body = response.errorBody();
                 if (body != null) {
@@ -163,7 +165,7 @@ public class InfluxException extends RuntimeException {
                         errorBody = new Gson().fromJson(json, new TypeToken<Map<String, Object>>() {
                         }.getType());
                         if (errorBody.containsKey("message")) {
-                            return errorBody.get("message").toString();
+                            return String.format(HTTP_STATUS_CODE_MESSAGE, code, errorBody.get("message").toString());
                         }
                     }
                 }
@@ -177,7 +179,7 @@ public class InfluxException extends RuntimeException {
                     .orElse(null);
 
             if (value != null) {
-                return value;
+                return String.format(HTTP_STATUS_CODE_MESSAGE, code, value);
             }
         }