+++ client/src/test/java/com/influxdb/client/ITVariablesApi.java
@@ -198,7 +198,7 @@ class ITVariablesApi extends AbstractITClientTest {
 
         Assertions.assertThatThrownBy(() -> variablesApi.findVariableByID("020f755c3c082000"))
                 .isInstanceOf(NotFoundException.class)
-                .hasMessage("variable not found for key \"020f755c3c082000\"");
+                .hasMessage("HTTP status code: 404; Message: variable not found for key \"020f755c3c082000\"");
     }
 
     @Test