+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 ## 6.3.0 [unreleased]
 
+### Features
+1. [#367](https://github.com/influxdata/influxdb-client-java/pull/367): Add HTTP status code to detail message of `InfluxException`
+1. [#367](https://github.com/influxdata/influxdb-client-java/pull/367): Add `GatewayTimeoutException` for HTTP status code 504
+
 ## 6.2.0 [2022-06-24]
 
 ### Features